Music페이지 추가 및 미디어쿼리(엉성함)

+++ src/page/member/modal.jsx
@@ -8,6 +8,19 @@ import { getYMDHM } from "utils/format-data";
 import logoUrl from "assets/img/logo/logo.png";
 import { useDispatch } from "react-redux";
 import { deleteLetter, modifyLetter } from "redux/modules/letter";
+import { TextShadow } from "components/text";
+
+const BackDrop = styled.div`
+  position: absolute;
+  top: 0;
+  bottom: 0;
+  left: 0;
+  right: 0;
+  background-color: var(--color-black);
+  overflow: hidden;
+  cursor: pointer;
+  z-index: 1;
+`;
 
 const Container = styled.section`
   width: 500px;
@@ -20,16 +33,27 @@ const Container = styled.section`
   transform: translate(-50%, -50%);
   display: flex;
   flex-direction: column;
+  z-index: 2;
+  border-radius: 8px;
+  box-shadow: 1px 2px 10px var(--color-white);
 `;
 
 const ModalTop = styled.div`
   display: flex;
   justify-content: space-between;
+  color: var(--color-black);
   align-items: center;
   p {
+    color: var(--color-black);
     white-space: nowrap;
     font-size: 0.8rem;
     font-weight: 600;
+  }
+  h1 {
+    color: var(--color-black);
+    font-size: var(--font-md);
+    font-weight: 600;
+    text-align: center;
   }
 `;
 const AvaterWrapper = styled.div`
@@ -41,37 +65,11 @@ const AvaterWrapper = styled.div`
     flex: 1;
     font-size: var(--font-mg);
     color: var(--color-black);
-    padding: var(--spacing) calc(var(--spacing) * 4);
-  }
-`;
-
-const ModalBottom = styled.div`
-  margin: calc(var(--spacing) * 4) 0;
-  flex: 1;
-
-  h1 {
+    padding: calc(var(--spacing) * 2);
     font-size: var(--font-md);
-    font-weight: bold;
-    padding: calc(var(--spacing) * 2) 0;
-  }
-  p {
-    line-height: 1.1;
+    font-weight: 600;
   }
 `;
-
-const ModalButtons = styled.div`
-  margin-left: auto;
-  display: flex;
-  gap: calc(var(--spacing) * 6);
-  button {
-    font-size: var(--font-md);
-    cursor: pointer;
-  }
-  svg {
-    fill: var(--color-primary-alt);
-  }
-`;
-
 const Avatar = styled.div`
   border-radius: 50%;
   width: 50px;
@@ -83,6 +81,27 @@ const Avatar = styled.div`
   background-image: ${(props) => `url(${props.$img})`};
 `;
 
+const ModalBottom = styled.div`
+  margin: calc(var(--spacing) * 4) 0;
+  flex: 1;
+  background-color: var(--color-dark-pink);
+  border-radius: 12px;
+  filter: brightness(105%);
+  h1 {
+    font-size: var(--font-lg);
+    font-weight: bold;
+    margin: 0 0 calc(var(--spacing) * 2) 0;
+    padding: calc(var(--spacing) * 2);
+  }
+  p {
+    line-height: 1.1;
+    color: var(--color-black);
+    padding-left: calc(var(--spacing) * 3);
+  }
+`;
+
+const TextStrong = styled(TextShadow)``;
+
 const ContentArea = styled.textarea`
   resize: none;
   color: black;
@@ -90,8 +109,27 @@ const ContentArea = styled.textarea`
   width: 100%;
   padding: calc(var(--spacing) * 2);
   line-height: 1.1;
+  font-size: 1.1rem;
   &:focus {
     outline: 3px solid var(--color-primary-alt);
+  }
+`;
+
+const ModalButtons = styled.div`
+  margin-left: auto;
+  display: flex;
+  gap: calc(var(--spacing) * 6);
+  button {
+    font-size: var(--font-md);
+    cursor: pointer;
+  }
+  button:hover {
+    svg {
+      fill: var(--color-accent);
+    }
+  }
+  svg {
+    fill: var(--color-primary-alt);
   }
 `;
 
@@ -101,20 +139,31 @@ export default function Modal({ selectedLetter, setLocalstorageLetters }) {
   const [contentValue, setContentValue] = useState(`${content}`);
   const [nickNameValue, setNickNameValue] = useState(nickname);
   const navigate = useNavigate();
-  const params = useParams();
+  const { name } = useParams();
   const dispatch = useDispatch();
 
   function onClickDelete() {
-    dispatch(deleteLetter(id));
-
-    navigate(`../${params.name}`);
+    const agreed = window.confirm("정말로 삭제하시겠습니까?");
+    if (agreed) {
+      dispatch(deleteLetter(id));
+      navigate(`../${name}`);
+    } else return;
   }
 
   function onClickSave() {
+    if (contentValue === content) {
+      return alert("수정된 내용이 없습니다.");
+    }
     dispatch(
       modifyLetter({ id, nickname: nickNameValue, content: contentValue, createdAt: getYMDHM(), writedTo, avatar })
     );
     setIsModify(false);
+  }
+
+  function onClickBackDrop() {
+    if (isModify) {
+      return alert("수정중입니다!.");
+    } else return navigate(`../${name}`);
   }
 
   useEffect(() => {
@@ -126,51 +175,55 @@ export default function Modal({ selectedLetter, setLocalstorageLetters }) {
   }, []);
 
   return (
-    <Container>
-      {isModify ? (
-        <>
-          <ModalTop>
-            <AvaterWrapper>
-              <Avatar $img={avatar} />
-              <input onChange={(e) => setNickNameValue(e.target.value)} value={nickNameValue} />
-            </AvaterWrapper>
-          </ModalTop>
-          <ModalBottom>
-            <ContentArea
-              draggable={false}
-              onChange={(e) => setContentValue(e.target.value)}
-              value={contentValue}
-            ></ContentArea>
-          </ModalBottom>
-          <ModalButtons>
-            <button onClick={onClickSave}>
-              <BiSolidSave />
-            </button>
-          </ModalButtons>
-        </>
-      ) : (
-        <>
-          <ModalTop>
-            <AvaterWrapper>
-              <Avatar $img={avatar} />
-              <h1>{nickname}</h1>
-            </AvaterWrapper>
-            <p>{createdAt}</p>
-          </ModalTop>
-          <ModalBottom>
-            <h1>To {writedTo}</h1>
-            <p>{content}</p>
-          </ModalBottom>
-          <ModalButtons>
-            <button onClick={() => setIsModify(true)}>
-              <BsFillPencilFill />
-            </button>
-            <button>
-              <RiDeleteBin6Fill onClick={onClickDelete} />
-            </button>
-          </ModalButtons>
-        </>
-      )}
-    </Container>
+    <>
+      <BackDrop onClick={onClickBackDrop} />
+      <Container>
+        {isModify ? (
+          <>
+            <ModalTop>
+              <AvaterWrapper>
+                <Avatar $img={avatar} />
+                <input maxLength={20} onChange={(e) => setNickNameValue(e.target.value)} value={nickNameValue} />
+              </AvaterWrapper>
+            </ModalTop>
+            <ModalBottom>
+              <ContentArea
+                maxLength={100}
+                draggable={false}
+                onChange={(e) => setContentValue(e.target.value)}
+                value={contentValue}
+              ></ContentArea>
+            </ModalBottom>
+            <ModalButtons>
+              <button onClick={onClickSave}>
+                <BiSolidSave />
+              </button>
+            </ModalButtons>
+          </>
+        ) : (
+          <>
+            <ModalTop>
+              <AvaterWrapper>
+                <Avatar $img={avatar} />
+                <h1>{nickname}</h1>
+              </AvaterWrapper>
+              <p>{createdAt}</p>
+            </ModalTop>
+            <ModalBottom>
+              <TextStrong>To {writedTo}</TextStrong>
+              <p>{content}</p>
+            </ModalBottom>
+            <ModalButtons>
+              <button onClick={() => setIsModify(true)}>
+                <BsFillPencilFill />
+              </button>
+              <button>
+                <RiDeleteBin6Fill onClick={onClickDelete} />
+              </button>
+            </ModalButtons>
+          </>
+        )}
+      </Container>
+    </>
   );
 }
